move bench-only function to benches
The jets_bench crate does not compile right now because it is not
included in the top-level workspace, and apparently it is not covered by
CI. Rather than fixing that here, I'll just throw this function into
there, adding to the compilation problems, and we'll fix it in a
followup PR.
